+++ programs/drift_vaults/src/instructions/initialize_tokenized_vault_depositor.rs
@@ -64,7 +64,7 @@ pub struct InitializeTokenizedVaultDepositor<'info> {
     pub vault: AccountLoader<'info, Vault>,
     #[account(
         init,
-        seeds = [b"tokenized_vault_depositor", vault.key().as_ref(), vault.load()?.shares_base.to_string().as_bytes().as_ref()],
+        seeds = [b"tokenized_vault_depositor", vault.key().as_ref(), vault.load()?.shares_base.to_string().as_bytes()],
         space = TokenizedVaultDepositor::SIZE,
         bump,
         payer = payer
@@ -72,7 +72,7 @@ pub struct InitializeTokenizedVaultDepositor<'info> {
     pub vault_depositor: AccountLoader<'info, TokenizedVaultDepositor>,
     #[account(
         init,
-        seeds = [b"mint", vault.key().as_ref(), vault.load()?.shares_base.to_string().as_bytes().as_ref()],
+        seeds = [b"mint", vault.key().as_ref(), vault.load()?.shares_base.to_string().as_bytes()],
         bump,
         payer = payer,
         mint::decimals = params.decimals,